Remove Agent RC staging deployment job (#44448)
### What does this PR do?

This PR changes the gitlab job for internal staging k8s Agent RC deployments. It removes the reference to the 'deployment' itself as this won't happen. Instead we will just trigger the child pipeline which will create the staging k8s deployment PR to handle manually. Because of that we just need to pass the new Agent image tag to the child pipeline, rest is not needed.

### Motivation

We make this job obsolete as these days we rely much more on the nightly runs and migrating the deployment part of this workflow out of gitlab is more work than benefit.

Co-authored-by: kacper.murzyn <kacper.murzyn@datadoghq.com>