chore: test framework deploy on ecs fargate only
The ecs scenario should be able to deploy the test workloads on fargate
nodes only.
For that: check if agent is enabled and if so deploy a fakeintake.
When checking for fargate deployment check only if test workload is
enabled and if fargate is provided, then deploy the fargate specific
workloads.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>