Add Check receiver to Interval method in `kubeletconfig.go` (#44683)
### What does this PR do?
The Interval method of `kubeletconfig.go` did not declare itself to be a method of `Check`, so the scheduler is using the default check interval of 15s instead of 10m. This means we're sending unnecessary data to the orchestrator pipeline which will inevitably get dropped by the RV cache. 

### Motivation
Bug spotted by @justin-lesko 

### Describe how you validated your changes
Build the agent, deployed on a local minikube cluster, and checked the logs:
`2025-12-30 21:45:49 UTC | CORE | INFO | (pkg/collector/scheduler/scheduler.go:93 in Enter) | Scheduling check orchestrator_kubelet_config:45ec97f32309a951 with an interval of 10m0s`

### Additional Notes


Co-authored-by: jon.rosario <jon.rosario@datadoghq.com>